Update IProviderManager documentation for similarity providers
