+++ tests/Jellyfin.Api.Tests/Controllers/SimilarityProvidersTests.cs
@@ -1,11 +1,8 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using MediaBrowser.Controller.Entities;
 using MediaBrowser.Controller.Entities.Audio;
-using MediaBrowser.Controller.Providers;
 using Xunit;
 
 namespace Jellyfin.Api.Tests.Controllers;
@@ -41,25 +38,5 @@ public class SimilarityProvidersTests
         // Assert
         Assert.NotNull(result);
         Assert.Empty(result);
-    }
-}
-
-public class MockSimilarityProvider : IItemSimilarityProvider<Audio>
-{
-    private readonly IEnumerable<Guid> _results;
-
-    public MockSimilarityProvider(IEnumerable<Guid> results)
-    {
-        _results = results;
-    }
-
-    public string Name => "MockProvider";
-
-    public Task<IEnumerable<Guid>> GetSimilarItems(
-        Audio item,
-        int limit,
-        CancellationToken cancellationToken)
-    {
-        return Task.FromResult(_results.Take(limit));
     }
 }
